Changed method of adding values to the hash table
Using TABLE[key] = value was causing many errors. Possible because the
corresponding item hadn't been constructed. I'm not sure. The copy
operation was failing. The "Other" in the copy constructor had a nullptr
for the package. This shouldn't have been possible, since we just
constructed it, and the assertion at the end of that constructor that
p!=nullptr passed. There must be some issue with the copy operation,
unless the entire copy operation was invalid, due to the value portion
of the table element not having been constructed yet... Which seems
ridiculous.